+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/send/SendViewTestSetup.kt
@@ -142,7 +142,8 @@ class SendViewTestSetup(
                             ) {
                                 onSettingsCount.incrementAndGet()
                             }
-                    )
+                    ),
+                isSyncing = false
             )
         }
     }